fix: rtc stream view extract
