+++ lib/features/call/widgets/widgets.dart
@@ -1,2 +1,3 @@
 export 'call_actions.dart';
 export 'call_info.dart';
+export 'rtc_stream_view.dart';
